Fix a bunch of places where we use longs where ints would do just fine.
Mostly because Gremlin uses longs to count things. Of course, of we ever
end up with more than Integer.MAX_VALUE of something the joke's on me.

+++ ehri-core/src/main/java/eu/ehri/project/models/DocumentaryUnit.java
@@ -120,7 +120,7 @@ public interface DocumentaryUnit extends AbstractUnit {
      */
     @Meta(CHILD_COUNT)
     @JavaHandler
-    long getChildCount();
+    int getChildCount();
 
     /**
      * Get child documentary units
@@ -152,8 +152,8 @@ public interface DocumentaryUnit extends AbstractUnit {
      */
     abstract class Impl implements JavaHandlerContext<Vertex>, DocumentaryUnit {
 
-        public long getChildCount() {
-            return gremlin().inE(Ontology.DOC_IS_CHILD_OF).count();
+        public int getChildCount() {
+            return Math.toIntExact(gremlin().inE(Ontology.DOC_IS_CHILD_OF).count());
         }
 
         public Iterable<DocumentaryUnit> getChildren() {
