+++ ehri-core/src/main/java/eu/ehri/project/models/base/ItemHolder.java
@@ -25,5 +25,5 @@ package eu.ehri.project.models.base;
 public interface ItemHolder {
     String CHILD_COUNT = "childCount";
 
-    long getChildCount();
+    int getChildCount();
 }
